// src/API.ts
export type Thread = {
  __typename: "Thread",
  id: string,
  topic?: string | null,
  comments?: ModelCommentConnection | null,
  createdAt?: string | null,
  updatedAt: string,
  owner?: string | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items:  Array<Comment | null >,
  nextToken?: string | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  owner?: string | null,
  body: string,
  thread: Thread,
  createdAt?: string | null,
  updatedAt: string,
  threadCommentsId?: string | null,
};
// ...
export type ModelThreadConnection = {
  __typename: "ModelThreadConnection",
  items:  Array<Thread | null >,
  nextToken?: string | null,
};
// ...
export type CreateThreadMutation = {
  createThread?:  {
    __typename: "Thread",
    id: string,
    topic?: string | null,
    comments?:  {
      __typename: "ModelCommentConnection",
      items:  Array< {
        __typename: "Comment",
        id: string,
        owner?: string | null,
        body: string,
        createdAt?: string | null,
        updatedAt: string,
        threadCommentsId?: string | null,
      } | null >,
      nextToken?: string | null,
    } | null,
    createdAt?: string | null,
    updatedAt: string,
    owner?: string | null,
  } | null,
};
// ...
export type ListThreadsQuery = {
  listThreads?:  {
    __typename: "ModelThreadConnection",
    items:  Array< {
      __typename: "Thread",
      id: string,
      topic?: string | null,
      comments?:  {
        __typename: "ModelCommentConnection",
        nextToken?: string | null,
      } | null,
      createdAt?: string | null,
      updatedAt: string,
      owner?: string | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};